Revert ItemFunctor.h/cc to its previous state, moving Variadic ParalleleFor to RangeFunctor
